PR feedback: move watchdog package to ee, rename docs to README.md
