+++ ee/watchdog/controller_windows.go
