+++ ee/watchdog/controller_windows_test.go
