+++ ee/watchdog/watchdog_service_windows.go
